Move sctSignatureAlg function
